Restrict access to K8s cluster in secgroups
The default Magnum settings allow all traffic from anywhere to minion
hosts in a Kubernetes cluster. This includes access to internal system
services like cAdvisor, which exposes information about internal system
state. What we would ideally want is that only services running in K8s
and the K8s API are exposed.

Restrict the security groups of cluster deployed by Magnum so that only
NodePort ports (30000-32767) are exposed to the outside world on the
minions by default. Also restrict access to the etcd ports on the master
node to the internal network. This means users will have to allow wider
access explicitly.