Enable OpenStack access for kube-controller-manager
Provide proper credentials and enable the openstack cloud provider when
launching the Kubernetes controller manager in a pod. Modify the
location of the system's CA directory that is mounted to match the
location of the actual CA bundle files and not the symbolic links under
/etc/ssl/certs. If only the symbolic links are accessible, then
Kubernetes won't be able to talk to endpoints secured by TLS because it
can't access the actual CA bundle files.